% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_color_vctr.R
\name{as_color_vctr}
\alias{as_color_vctr}
\title{Coerce object to a color_vctr}
\usage{
as_color_vctr(x, text_color = NA, background = NA, style = NA)
}
\arguments{
\item{x}{object to coerce}

\item{text_color}{A vector of length 1 or same length as vect. Details the
color the text should be. Valid values can be found from the
`valid_text_color()` function.NA means no text color.}

\item{background}{A vector of length 1 or same length as vect. Details the
background color of the text. Valid values can be found from the
`valid_background()` function. NA means no background color.}

\item{style}{A vector of length 1 or same length as vect. Details the
style of the text Valid values can be found from the
`valid_style()` function. NA means no styling.}
}
\description{
Coersion function for generation of a color_vctr from an existing vector
}
