% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortable_vect.R
\name{c_vctr}
\alias{c_vctr}
\title{vector of colortable_cells}
\usage{
c_vctr(x, ..., text_color = NA, background = NA, style = NA)
}
\arguments{
\item{...}{colortable_cells}
}
\value{
a vector of colortable_cells
}
\description{
`c` destroys the attributes of the contents, c_vctr preserves them
}
