% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color2hex.R
\name{color2hex}
\alias{color2hex}
\title{Replaces color names with hex-color names}
\usage{
color2hex(x)
}
\arguments{
\item{x}{a vector, matrix, or data.frame containing color names}
}
\value{
hex-color-names in the same structur as x
}
\description{
Replaces color names with hex-color names
}
\examples{
color2hex("red")
color2hex("#ff0000")

color2hex(c("red", "green"))

color2hex(matrix(c("red", "green", "blue", "yellow"), 2, 2))

color2hex(data.frame(x = c("red", "green"),
                     y = c("blue", "yellow")))
}
