% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_gradient.R
\name{color_gradient}
\alias{color_gradient}
\title{Creates a color gradient for a given vector}
\usage{
color_gradient(x, colors = c("green", "yellow", "red"), x_range = NA,
  na_color = "gray", colsteps = 100)
}
\arguments{
\item{x}{a vector of numeric values}

\item{colors}{the colors for the gradient}

\item{x_range}{a vector of length 2 that specifies the range of x, defaults to min/max of x}

\item{na_color}{the default color if a point is outside of the x_range, defaults to gray}

\item{colsteps}{the number of color-steps, defaults to 100}
}
\value{
a vector of colors of the same length as x
}
\description{
Creates a color gradient for a given vector
}
\examples{
x <- c(5, 1, 2, 5, 4, 3)
color_gradient(x, c("red", "green"))
color_gradient(x, c("red", "green"), x_range = c(0, 10))

color_gradient(x, c("red", "black", "green"))

x <- seq(-10, 10, length.out = 100)
y <- abs(x^2 - 20)
plot(x, y, col = color_gradient(y), pch = 19, cex = 2)
plot(x, y, col = color_gradient(y, c("yellow", "blue")), pch = 19, cex = 2)
plot(x, y, col = color_gradient(y, x_range = c(10, 20)), pch = 19, cex = 2)

}
