% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils_key.R
\docType{data}
\name{color_key}
\alias{color_key}
\title{return the tibble of color keys}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 256 rows and 5 columns.}
\usage{
color_key
}
\description{
Source is from https://jonasjacek.github.io/colors/
}
\keyword{datasets}
