% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_console.R
\docType{data}
\name{color_key_console}
\alias{color_key_console}
\title{return the tibble of color keys}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 256 rows and 4 columns.}
\usage{
color_key_console
}
\description{
Source is from https://jonasjacek.github.io/colors/
}
\keyword{datasets}
