% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_html.R
\docType{data}
\name{color_key_html}
\alias{color_key_html}
\title{return the tibble of valid web colors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 148 rows and 3 columns.
}
\usage{
color_key_html
}
\description{
Source is from https://www.w3schools.com/colors/color_tryit.asp
}
\keyword{datasets}
