% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_tex.R
\docType{data}
\name{color_key_latex}
\alias{color_key_latex}
\title{return the tibble of valid web colors}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 718 rows and 4 columns.}
\usage{
color_key_latex
}
\description{
Source is from https://www.w3schools.com/colors/color_tryit.asp
}
\keyword{datasets}
