% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_vctr.R
\name{color_scale}
\alias{color_scale}
\title{scale color vector colors based on contents}
\usage{
color_scale(palette, na.color = "#808080")
}
\arguments{
\item{palette}{pallete to use. A vector of colors to use}

\item{na.color}{color to set NA values}
}
\description{
scale color vector colors based on contents
}
