% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_table.R
\name{color_table}
\alias{color_table}
\title{Creates a html or latex colored table}
\usage{
color_table(x, body_bg = "#74add1", header_bg = "dodgerblue2",
  body_color = "black", body_bold = FALSE, header_color = "black",
  header_bold = TRUE, align = "c", hide_body = FALSE, row_bgs = NA,
  col_bgs = NA, view = FALSE, cell_width = "15px", format)
}
\arguments{
\item{x}{a matrix or data.frame to be displayed}

\item{body_bg}{a single value or a matrix/data.frame of the same size of x
with the background colors for the body of the table}

\item{header_bg}{a single value or a vector of the same size as the number of
columns of x, containing the background color(s) for the header}

\item{body_color}{a single value or a matrix/data.frame of the same size of
x with the text colors for the body of the table, if NA (default) takes the
same value as the background color}

\item{body_bold}{T/F if the body-text is bold}

\item{header_color}{a single value or a vector of the same size as the number
of columns of x, containing the text-color(s) for the header}

\item{header_bold}{T/F if the header-text is bold}

\item{align}{The alignment of the columns}

\item{hide_body}{if the contents of the body of x are hidden
(replaced with "")}

\item{row_bgs}{a vector of colors for the rows (will be extended if
necessary)}

\item{col_bgs}{a vector of colors for the columns (will be extended if
necessary)}

\item{view}{T/F when the format is html, should it be displayed directly?}

\item{cell_width}{a minimum width of the columns (applies to html only),
defaults to 15px}

\item{format}{a format for the code, is passed to \code{knitr::kable},
allowed is either html or latex}
}
\value{
the latex/html code (invisible in case of html-output and view = T)
}
\description{
Creates a html or latex colored table
}
\examples{
x <- data.frame(x = 1:3, y = 2:4, z = 3:5)
color_table(x, format = "html", view = T)

# remove the text color of the body
color_table(x, body_color = NA, format = "html", view = T)
color_table(x, body_color = NA, header_color = NA, format = "html", view = T)

# remove the text color of the header
color_table(x, body_color = NA, header_color = NA, format = "html", view = T)

# change the colors:
color_table(x, body_color = "red", header_color = "white", format = "html",
            view = T)

color_table(x, body_color = "red", header_color = c("red", "green", "blue"),
            format = "html", view = T)

# color columns:
color_table(x, col_bgs = c("red", "green", "blue"), format = "html", view = T)

# color rows:
color_table(x, row_bgs = c("red", "green", "blue"), format = "html", view = T)
# can be subsets:
large_x <- data.frame(x = 1:9, y = 1:9, z = 1:9)
color_table(large_x, row_bgs = c("blue", "green"), format = "html", view = T)


# color cells individually
colors1 <- matrix(c("red", "green", "blue", "yellow", "purple", "white",
                    "black", "gray", "yellowgreen"),
                  nrow = nrow(x), ncol = ncol(x), byrow = T)

color_table(x, body_bg = colors1, format = "html", view = T)

# special case: body_bg/color contains the header-information as well
colors2 <- matrix(c("red", "red", "red",
                    "green", "blue", "green",
                    "blue", "green", "blue",
                    "green", "blue", "green"),
                  ncol = 3, byrow = T)
color_table(x, body_bg = colors2, format = "html", view = T)

# specify widths of the cell
x2 <- data.frame(x = 1:10, x2 = 2^(1:10))
color_table(x2, format = "html", view = T) -> a

color_table(x2, cell_width = "75px", format = "html", view = T)
}
