% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortable_vect.R
\name{colortable_vect}
\alias{colortable_vect}
\title{vector of colortable_cells}
\usage{
colortable_vect(x, ..., text_color = NA, background = NA, style = NA)
}
\arguments{
\item{...}{colortable_cells}
}
\value{
a vector of colortable_cells
}
\description{
`c` destroys the attributes of the contents, colortable_vect preserves them
}
