% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_color_vctr.R
\name{escape_chars}
\alias{escape_chars}
\title{Protect control characters in a string for use in a latex table or caption}
\usage{
escape_chars(x)
}
\arguments{
\item{x}{character vector of test containing values that need to be latex escaped}
}
\value{
character Vector of transformed values for table output
}
\description{
escape takes a vector of chracter values, and puts "\\" in front of them to make them
allowable in the latex table output
}
\examples{
value_example <- c("testvalue", "test_value", "ampersand&")
escape_chars(value_example)
escape_chars("String_Entry \%")

}
