% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{field}
\alias{field}
\title{Access Attributes consistent with vctrs}
\usage{
field(x, i)
}
\arguments{
\item{x}{color_vctr object}

\item{i}{attribute to access}
}
\value{
specified attribute
}
\description{
provide a utility to access contents of colorvctrs consistent with vctrs
}
