% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_colortable_cell.R
\name{format.colortable_cell}
\alias{format.colortable_cell}
\title{format method for colortable_cells}
\usage{
\method{format}{colortable_cell}(x, ..., method = print_method())
}
\arguments{
\item{x}{object of colortable cell}
}
\description{
format method for colortable_cells
}
