% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_colortable_vect.R
\name{format.colortable_vect}
\alias{format.colortable_vect}
\title{format method for colortable_vector}
\usage{
\method{format}{colortable_vect}(x, ..., method = print_method())
}
\arguments{
\item{x}{object of colortable cell}
}
\description{
format method for colortable_vector
}
