% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vctr.R
\name{is_color_vctr}
\alias{is_color_vctr}
\title{Is the object a color_vctr?
s
Detect if object is a colortable vector}
\usage{
is_color_vctr(x)
}
\arguments{
\item{x}{object to be checked if is a color_vctr}
}
\description{
Is the object a color_vctr?
s
Detect if object is a colortable vector
}
