% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.color_vctr}
\alias{knit_print.color_vctr}
\title{knit_print method for color_vctrs}
\usage{
\method{knit_print}{color_vctr}(x, inline = FALSE, ...)
}
\description{
knit_print method for color_vctrs
}
\keyword{internal}
