% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.data.frame}
\alias{knit_print.data.frame}
\title{knit_print method for data.frames}
\usage{
\method{knit_print}{data.frame}(x, options, ...)
}
\description{
knit_print method for data.frames
}
\keyword{internal}
