% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vctr_arith.R
\name{merge_styling}
\alias{merge_styling}
\title{Utilities to merge styling of two vectors}
\usage{
merge_styling(x, y)
}
\arguments{
\item{x}{color_vctr}

\item{y}{color_vctr}
}
\description{
Utilities to merge styling of two vectors
}
