% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vctr.R
\name{new_color_vctr}
\alias{new_color_vctr}
\title{Create a color_vctr from source data}
\usage{
new_color_vctr(
  vect = double(),
  text_color = NA_character_,
  background = NA_character_,
  style = NA_character_
)
}
\arguments{
\item{vect}{The input vector to convert to a color_vctr. Must be an atomic.}

\item{text_color}{A vector of length 1 or same length as vect. Details the
color the text should be. Valid values can be found from the
`valid_colors()` function.NA means no text color.}

\item{background}{A vector of length 1 or same length as vect. Details the
background color of the text. Valid values can be found from the
`valid_colors()` function. NA means no background color.}

\item{style}{A vector of length 1 or same length as vect. Details the
style of the text Valid values can be found from the
`valid_style()` function. NA means no styling.}
}
\description{
This is a method to create the a new color_vctr class
}
