% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_html.R
\name{postprocess_html}
\alias{postprocess_html}
\title{postprocesses the html-code}
\usage{
postprocess_html(x, width = "15px")
}
\arguments{
\item{x}{a structure containing the html code that can be coerced to
characters}

\item{width}{the width of the cells, defaults to 15px}
}
\value{
a knitr_kable object
}
\description{
The goal is to exclude the span and shift the styles to td/th
}
\details{
This makes sure that the whole box is colored instead of only the values
}
\examples{
# internally used only
}
\keyword{internal}
