% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_c_vctr.R
\name{print.color_vctr}
\alias{print.color_vctr}
\title{print method for color_vctr}
\usage{
\method{print}{color_vctr}(x, ..., method = print_method())
}
\arguments{
\item{x}{object of colortable vector}

\item{...}{print options to be passed on}

\item{method}{The output type to print to. Defaults to one of: "console","latex","html".}
}
\description{
print method for color_vctr
}
