% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_colortable_cell.R
\name{print.colortable_cell}
\alias{print.colortable_cell}
\title{print method for colortable_cells}
\usage{
\method{print}{colortable_cell}(x, ..., method = print_method())
}
\arguments{
\item{x}{object of colortable cell}
}
\description{
print method for colortable_cells
}
