% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_colortable_vect.R
\name{print.colortable_vect}
\alias{print.colortable_vect}
\title{print method for colortable_vect}
\usage{
\method{print}{colortable_vect}(x, ..., method = print_method())
}
\arguments{
\item{x}{object of colortable vector}

\item{...}{print options to be passed on}

\item{method}{The output type to print to. Defaults to one of: "console","latex","html".}
}
\description{
print method for colortable_vect
}
