% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_frame_color_vector.R
\name{print.data.frame}
\alias{print.data.frame}
\title{print data.frame}
\usage{
\method{print}{data.frame}(
  x,
  ...,
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = TRUE,
  max = NULL
)
}
\description{
custom print method for data.frames that contain color_vectors
}
