% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_method}
\alias{print_method}
\title{Determine Printing for colortable}
\usage{
print_method()
}
\value{
character vector
}
\description{
determines if table is being printed interactively, or in an rmd
}
