% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_vctr.R
\name{scale_color}
\alias{scale_color}
\title{scale color vector colors based on contents}
\usage{
scale_color(palette, na.color = "#808080")
}
\arguments{
\item{palette}{pallete to use. A vector of colors to use}
}
\description{
color vector contents based on contents. only to be used within
color_vctr with numeric values
}
