% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000.R
\name{set_color_vctr_precedence}
\alias{set_color_vctr_precedence}
\title{Merge semantics for colortable vector arithmetic}
\usage{
set_color_vctr_precedence(precedence = c("left", "right", "blended", "mixed"))
}
\arguments{
\item{precedence}{one of three options: left, right or blended}
}
\description{
Set how tables with different styling should merge,
by deffering to left styling when it exists, right styling,
or attempt to blend them together.
}
