% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_styling.R
\name{set_styling}
\alias{set_styling}
\title{Update/Set styling of colortable elements}
\usage{
set_styling(
  x,
  idx = rep(TRUE, length(x)),
  text_color = NA,
  background = NA,
  style = NA
)
}
\arguments{
\item{x}{input vector to be styled}

\item{idx}{the indexes to be updated or styled. Either numeric or logical}

\item{text_color}{a valid text color to set the elements}

\item{background}{a valid background color to set the elements}

\item{style}{a valid style to apply to the elements}
}
\value{
colortable_vect with applied/updated styling
}
\description{
Update/Set styling of colortable elements
}
