% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{style2ansi}
\alias{style2ansi}
\title{Create function to append ansi coloring}
\usage{
style2ansi(x, style = NA, text_color = NA, background = NA, ...)
}
\arguments{
\item{color}{the color wanted. must be onen of the colors in the color_codes list}
}
\value{
function to append colors
}
\description{
Create function to append ansi coloring
}
\examples{
colored_text <- colortable:::style2ansi(24, text_color = "red" )
cat(colored_text)

}
