% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_entries.R
\name{Valid}
\alias{Valid}
\alias{valid_text_color}
\alias{valid_background}
\alias{valid_style}
\title{Valid Coloring/Styling Options for {{Colortable}}}
\usage{
valid_text_color
valid_background
valid_style

valid_text_color(method = c("latex", "html", "console"))

valid_background(method = c("latex", "html", "console"))

valid_style(method = c("latex", "html", "console"))
}
\arguments{
\item{method}{The output method type. One of "latex","html", or "console"}
}
\value{
vector of valid settings
}
\description{
Valid Coloring/Styling Options for {{Colortable}}

Return vector of valid text color options

Return vector of valid background color options

Return vector of valid text styling options
}
