% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vctr.R, R/color_vctr_arith.R
\name{vec_arith.color_vctr}
\alias{vec_arith.color_vctr}
\title{Arithmatic}
\usage{
\method{vec_arith}{color_vctr}(op, x, y, ...)

\method{vec_arith}{color_vctr}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Arithmatic

Arithmatic
}
