% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_html.R
\name{view_html}
\alias{view_html}
\title{Views HTML-code in a browser}
\usage{
view_html(x)
}
\arguments{
\item{x}{a string of html-code}
}
\description{
Views HTML-code in a browser
}
\examples{
view_html("<html><body>Hello there!</body></html>")
}
