% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vctr.R
\name{color_vctr}
\alias{color_vctr}
\title{Create a color_vctr}
\usage{
color_vctr(x = double(), text_color = NA, background = NA, style = NA)
}
\arguments{
\item{x}{data source determining method dispatch}

\item{text_color}{A vector of length 1 or same length as vect. Details the
color the text should be. Valid values can be found from the
`valid_text_color()` function.NA means no text color.}

\item{background}{A vector of length 1 or same length as vect. Details the
background color of the text. Valid values can be found from the
`valid_background()` function. NA means no background color.}

\item{style}{A vector of length 1 or same length as vect. Details the
style of the text Valid values can be found from the
`valid_style()` function. NA means no styling.}
}
\value{
a color_vctr
}
\description{
This is the generic method method dispatches color_vctr generation based on
the first argument.
}
