% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_frame_color_vector.R
\name{print.data.frame}
\alias{print.data.frame}
\title{wrapper to print data.frame colors}
\usage{
\method{print}{data.frame}(
  x,
  ...,
  digits = NULL,
  quote = FALSE,
  right = TRUE,
  row.names = TRUE,
  max = NULL
)
}
\arguments{
\item{x}{a data.frame}

\item{...}{optional arguments to print or plot methods.}

\item{digits}{the minimum number of significant digits to be used: see print.default.}

\item{quote}{logical, indicating whether or not entries should be printed with surrounding quotes.}

\item{right}{logical, indicating whether or not strings should be right-aligned. The default is right-alignment.}

\item{row.names}{logical (or character vector), indicating whether (or what) row names should be printed.}

\item{max}{numeric or NULL, specifying the maximal number of entries to be printed. By default, when NULL, getOption("max.print") used.}
}
\description{
custom print method for data.frames that contain color_vctr's. Uses the
base print.data.frame otherwise
}
