% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vctr.R, R/print_color_maths.R
\name{vec_arith.color_vctr}
\alias{vec_arith.color_vctr}
\title{Arithmatic}
\usage{
\method{vec_arith}{color_vctr}(op, x, y, ...)

\method{vec_arith}{color_vctr}(op, x, y, ...)
}
\description{
Arithmatic

Arithmatic
}
